+++ apps/api/src/headers/responses/CORS_HEADERS.ts
@@ -1,9 +1,9 @@
 const CORS_HEADERS = {
-    "access-control-allow-credentials": true,
-    "access-control-allow-headers": "accept,authorization",
-    "access-control-allow-methods": "GET,HEAD,OPTIONS",
+    "access-control-allow-credentials": false,
+    "access-control-allow-headers": "accept,authorization,content-type",
+    "access-control-allow-methods": "GET,HEAD,OPTIONS,POST",
     "access-control-allow-origin": "*",
-    "access-control-expose-headers": "*,authorization",
+    "access-control-expose-headers": "*,authorization,location",
     "access-control-max-age": 7200, // capped in Chromium v76+
 }
 export default CORS_HEADERS